code clean up & readme written

+++ src/pages/home.page.jsx
@@ -1,8 +1,13 @@
 import React from 'react';
 import { withRouter, Link } from 'react-router-dom';
 import PropTypes from "prop-types";
-import './home.style.css';
-import { createDoc, getDocList, checkDocNameExist, deleteDoc } from '../../utility/document.service';
+import { createDoc, getDocList, checkDocNameExist, deleteDoc } from '../utility/document.service';
+import { cssClass as css} from '../utility/utility';
+
+const BarScroller= {
+    overflow: 'auto',
+    height: '580px'
+}
 
 class Home extends React.Component {
     constructor(props){
@@ -10,11 +15,11 @@ class Home extends React.Component {
         this.state={
             docList: getDocList(),
             docName: "",
-    
         }
         this.gotoBuidler = this.gotoBuidler.bind(this);
         this.handleInput= this.handleInput.bind(this);
         this.deleteItem = this.deleteItem.bind(this);
+        
     }
     static propTypes = {
         match: PropTypes.object.isRequired,
@@ -43,6 +48,7 @@ class Home extends React.Component {
     }
 
     deleteItem(id){
+        if(!window.confirm("Are you sure to delete?")){ return}
         deleteDoc(id);
         let newList = getDocList();
         this.setState({
@@ -60,10 +66,10 @@ class Home extends React.Component {
         return (
             <div className='h-full flex content-stretch'>
                 <div className='w-2/3 p-3 flex items-center justify-center'>
-                    <div className='shadow p-10 w-2/3 flex flex-col'>
+                    <div className='shadow p-10 w-2/3 flex border flex-col'>
                         <input onChange={this.handleInput} placeholder='Type the name of the doc' type="text" className='shadow-lg p-5  m-3 border text-lg'/>
                         <div className='flex justify-center'>
-                         <button className='bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded'
+                         <button className={css['btn-primay']}
                                onClick={this.gotoBuidler}
                                >
                                 Go
@@ -78,17 +84,17 @@ class Home extends React.Component {
                         Saved items
                     </div>
                     <div className='p-3' >
-                        <ul className='divide-y'>
+                        <ul style={BarScroller} className='divide-y'>
                             {docList.map(doc => {
                                 return (
-                                <li key={doc.docId} className='py-3 flex justify-between'>
+                                <li key={doc.docId} className='py-3 flex items-center justify-between'>
                                     <div className='w-2/3 flex items-center'>
                                         {doc.docName}
                                     </div>
-                                    <div className='w-1/3 flex justify-around'>
-                                        <Link to={`/preview/${doc.docId}`}>Preview</Link>
-                                        <Link to={`/builder/${doc.docId}`}>Edit</Link>
-                                        <button onClick={()=> this.deleteItem(doc.docId)}>Delete</button>
+                                    <div className='w-1/3 flex text-blue-400 items-center justify-around'>
+                                        <Link className='text-sm hover:bg-gray-300 p-1 rounded' to={`/preview/${doc.docId}`}>Preview</Link>
+                                        <Link className='text-sm hover:bg-gray-300 p-1 rounded' to={`/builder/${doc.docId}`}>Edit</Link>
+                                        <button className={css['btn-danger-sm']} onClick={()=> this.deleteItem(doc.docId)}>Delete</button>
                                     </div>
                                 </li>
                                 )
